Broke up old UIManager into UILayout and UIManager
- UILayout defines widgets and how they are placed
- UIManager provides an API to work with a layout
